+++ tests/corpus/contracts/panic_with_contracterror/src/lib.rs
@@ -2,7 +2,7 @@
 use soroban_sdk::{contract, contracterror, contractimpl, panic_with_error, Env};
 
 #[contracterror]
-#[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord)]
+#[derive(Copy, Clone)]
 #[repr(u32)]
 pub enum Error {
     Bad = 1,
@@ -14,6 +14,6 @@ pub struct Contract;
 #[contractimpl]
 impl Contract {
     pub fn safe_error(env: Env) {
-        panic_with_error!(env, Error::Bad);
+        panic_with_error!(&env, Error::Bad);
     }
 }